rate-limiter/requirements.txt
asyncio-nats-client==0.10.0
asyncio-nats-streaming==0.4.0
protobuf==3.12.2
python-dotenv==0.13.0
six==1.15.0
redis==3.0.1

